+++ index.html
@@ -4,7 +4,7 @@
   <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
+  <title>Awesome Books</title>
   <link rel="stylesheet" href="index.css">
 </head>
 <body>
@@ -58,6 +58,6 @@
   </main>
   <footer>Copyright - All rights reserved</footer>
   
-  <script src="index.js"></script>
+  <script type="module" src="index.js"></script>
 </body>
 </html>
